feat(pbo): Finish PboBuilder Up
Finished PboBuilder
Right now you cannot use the PboBuilder with a pre-existing pbo. Support for this should be added in one of the following commits.
Added enum for specifying how you want to create a pbo object.
Made some small fixes to the PboVersionEntry.cs class aswell as the main PboFile.cs