feat(RVBank): Add System For Handling Duplicate Entries
closes #gh-017

This commit introduces a system in RVBank for handling duplicate entries.

The changes include the addition of logging statements for better debugging and tracking of the progress of the process. The method 'InitializeData' in RVBankDataEntry.cs now returns a boolean value to check for successful initialization. Change has been made in 'ReadPbo' method of RVBank to enable more detailed tracking of the debinarization progress.

Handling duplicate entries are important to avoid data redundancy and ensure smooth operations. This feature will identify any duplicate entries and remove them, ensuring the integrity of the data. But in some cases, these duplicates will simply be ignored based on 'IgnoreDuplicateFiles' option. Also, at the end of the debinarization process, all deleted or stale entries will be removed from the directory structure.

Moreover, a small change is made in '.idea/workspace.xml' file to reflect the updated Git branch ('feature/duplicate-bank-entries') in the workspace

Signed-off-by: ryann <flipperplz@protonmail.com>